% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A01_methodObj.R, R/A03_methodObjSolver.R
\docType{methods}
\name{internal-fit}
\alias{internal-fit}
\alias{.fit,methodObj-method}
\alias{.predict,methodObj-method}
\alias{.fit,methodObjSolver-method}
\title{Execute Regression/Prediction Method}
\usage{
.fit(object, ...)

\S4method{.fit}{methodObj}(object, data, response, ...)

\S4method{.predict}{methodObj}(object, data, response, ...)

\S4method{.fit}{methodObjSolver}(object, ...)
}
\arguments{
\item{object}{A methodObj object}
}
\value{
The value object as defined by the regression/prediction method 
  called.
}
\description{
Execute regression/prediction method
}
\keyword{internal}
