% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A01_methodObj.R, R/A02_methodObjPredict.R
\docType{methods}
\name{internal-predict}
\alias{internal-predict}
\alias{.predict,methodObjPredict-method}
\title{Obtain predictions}
\usage{
.predict(object, ...)

\S4method{.predict}{methodObjPredict}(object, newdata, fitObj)
}
\arguments{
\item{object}{A methodObjPredict object}

\item{newdata}{A data.frame of model covariates}

\item{fitObj}{The value object returned by the regression}
}
\value{
The value object as defined by the prediction method called. If a
  vector is returned by method it is converted to a single column matrix.
}
\description{
Execute prediction method
}
\keyword{internal}
