% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A01_methodObj.R
\docType{methods}
\name{method}
\alias{method}
\alias{method,methodObj-method}
\title{Retrieve the method}
\usage{
method(object, ...)

\S4method{method}{methodObj}(object, ...)
}
\arguments{
\item{object}{An object of class \code{methodObj}}
}
\value{
A character or function.
}
\description{
Returns the character or function of the regression/prediction method
}
\keyword{internal}
