% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A01_methodObj.R
\docType{methods}
\name{methodArgs}
\alias{methodArgs}
\alias{methodArgs,methodObj-method}
\alias{methodArgs<-}
\alias{methodArgs<-,methodObj,list-method}
\title{Retrieve the arguments to be passed to method}
\usage{
methodArgs(object, ...)

\S4method{methodArgs}{methodObj}(object, ...)

methodArgs(object) <- value

\S4method{methodArgs}{methodObj,list}(object) <- value
}
\arguments{
\item{object}{An object of class \code{methodObj}}
}
\value{
A list.
}
\description{
Returns/Sets the list of arguments to be passed to the 
  regression/prediction method.
}
\keyword{internal}
