% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A01_methodObj.R
\docType{class}
\name{methodObj-class}
\alias{methodObj-class}
\title{Class \code{methodObj}}
\description{
Class \code{methodObj} contains a regression or prediction method and any
  user specified inputs to be passed to the regression or prediction method.
}
\section{Slots}{

\describe{
\item{\code{method}}{ANY A character name or function.}

\item{\code{methodArgs}}{A list of inputs to be passed to the method.}
}}

\keyword{internal}
