% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A02_methodObjPredict.R
\docType{class}
\name{methodObjPredict-class}
\alias{methodObjPredict-class}
\title{Class \code{methodObjPredict}}
\description{
Extends class \code{methodObj} to indicate that the method is a prediction
  method.
}
\section{Slots}{

\describe{
\item{\code{newdataName}}{A character giving the formal argument for input data.frame}

\item{\code{objectName}}{A character giving the formal argument for the input 
regression object}

\item{\code{propenMissing}}{A character indicating if a treatment variable is missing}

\item{\code{method}}{ANY A character name or function.}

\item{\code{methodArgs}}{A list of inputs to be passed to the method.}
}}

\keyword{internal}
