% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A03_methodObjSolver.R
\docType{class}
\name{methodObjSolver-class}
\alias{methodObjSolver-class}
\title{Class \code{methodObjSolver}}
\description{
Extends class \code{methodObj} to indicate that the method is a regression
  method.
}
\section{Slots}{

\describe{
\item{\code{method}}{ANY A character name or function.}

\item{\code{methodArgs}}{A list of inputs to be passed to the method.}
}}

\keyword{internal}
