% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A04_methodObjSolverFormula.R
\docType{class}
\name{methodObjSolverFormula-class}
\alias{methodObjSolverFormula-class}
\title{Class \code{methodObjSolverFormula}}
\description{
Extends class \code{methodObjSolver} to indicate formula input
}
\section{Slots}{

\describe{
\item{\code{dataName}}{A character giving the formal argument for input data.frame}

\item{\code{formulaName}}{A character giving the formal argument for the model input}

\item{\code{method}}{ANY A character name or function.}

\item{\code{methodArgs}}{A list of inputs to be passed to the method.}
}}

\keyword{internal}
