% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A05_methodObjSolverXY.R
\docType{class}
\name{methodObjSolverXY-class}
\alias{methodObjSolverXY-class}
\title{Class \code{methodObjSolverXY}}
\description{
Extends class \code{methodObjSolver} to indicate design matrix input
}
\section{Slots}{

\describe{
\item{\code{xName}}{A character giving the formal argument for design input}

\item{\code{yName}}{A character giving the formal argument for response input}

\item{\code{method}}{ANY A character name or function.}

\item{\code{methodArgs}}{A list of inputs to be passed to the method.}
}}

\keyword{internal}
