% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelObj.R, R/modelObjFit.R
\name{model}
\alias{model}
\alias{model,ANY-method}
\alias{model,modelObj-method}
\alias{model,modelObjFit-method}
\title{Retrieve model}
\usage{
model(object, ...)

\S4method{model}{ANY}(object, ...)

\S4method{model}{modelObj}(object, ...)

\S4method{model}{modelObjFit}(object, ...)
}
\arguments{
\item{object}{A modelObj object}

\item{...}{ignored}
}
\value{
The formula for the regression
}
\description{
Retrieves model from modelObj
}
