% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A01_methodObj.R, R/A02_methodObjPredict.R,
%   R/A03_methodObjSolver.R, R/A04_methodObjSolverFormula.R,
%   R/A05_methodObjSolverXY.R, R/A06_modelObj.R, R/A09_modelObjFit.R
\docType{methods}
\name{modelObj-internal-api}
\alias{modelObj-internal-api}
\alias{tmp}
\alias{.newMethodObjPredict,ANY-method}
\alias{.newMethodObjPredict,NULL-method}
\alias{.newMethodObjPredict,list-method}
\alias{.newMethodObjSolver,ANY-method}
\alias{.newMethodObjSolver,list-method}
\alias{.fit,methodObjSolverFormula-method}
\alias{.newMethodObjSolverFormula,ANY-method}
\alias{.newMethodObjSolverFormula,NULL-method}
\alias{.newMethodObjSolverFormula,list-method}
\alias{.fit,methodObjSolverXY-method}
\alias{.newMethodObjSolverXY,ANY-method}
\alias{.newMethodObjSolverXY,NULL-method}
\alias{.newMethodObjSolverXY,list-method}
\alias{model,modelObj-method}
\alias{solver,modelObj-method}
\alias{solverArgs,modelObj-method}
\alias{solverArgs<-,modelObj-method}
\alias{predictor,modelObj-method}
\alias{predictorArgs,modelObj-method}
\alias{predictorArgs<-,modelObj-method}
\alias{coef,modelObjFit-method}
\alias{plot,modelObjFit-method}
\alias{print,modelObjFit-method}
\alias{residuals,modelObjFit-method}
\alias{show,modelObjFit-method}
\alias{summary,modelObjFit-method}
\title{Hidden methods}
\usage{
tmp(x)

\S4method{.newMethodObjPredict}{ANY}(args, method)

\S4method{.newMethodObjPredict}{`NULL`}(args, method)

\S4method{.newMethodObjPredict}{list}(args, method)

\S4method{.newMethodObjSolver}{ANY}(args, method)

\S4method{.newMethodObjSolver}{list}(args, method)

\S4method{.fit}{methodObjSolverFormula}(object, data, response, model)

\S4method{.newMethodObjSolverFormula}{ANY}(args, method)

\S4method{.newMethodObjSolverFormula}{`NULL`}(args, method)

\S4method{.newMethodObjSolverFormula}{list}(args, method)

\S4method{.fit}{methodObjSolverXY}(object, data, response, model)

\S4method{.newMethodObjSolverXY}{ANY}(args, method)

\S4method{.newMethodObjSolverXY}{`NULL`}(args, method)

\S4method{.newMethodObjSolverXY}{list}(args, method)

\S4method{model}{modelObj}(object, ...)

\S4method{solver}{modelObj}(object, ...)

\S4method{solverArgs}{modelObj}(object, ...)

\S4method{solverArgs}{modelObj}(object) <- value

\S4method{predictor}{modelObj}(object, ...)

\S4method{predictorArgs}{modelObj}(object, ...)

\S4method{predictorArgs}{modelObj}(object) <- value

\S4method{coef}{modelObjFit}(object, ...)

\S4method{plot}{modelObjFit}(x, y, ...)

\S4method{print}{modelObjFit}(x)

\S4method{residuals}{modelObjFit}(object, ...)

\S4method{show}{modelObjFit}(object)

\S4method{summary}{modelObjFit}(object, ...)
}
\description{
Hidden methods
}
\keyword{internal}
