% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A09_modelObjFit.R
\docType{class}
\name{modelObjFit}
\alias{modelObjFit}
\title{Class \code{modelObjFit}}
\description{
A class for storing regression analysis results.
}
\section{Slots}{

\describe{
\item{\code{fitObj}}{Object returned by the regression analysis}

\item{\code{model}}{Object of class \code{formula}}

\item{\code{func}}{Object of class \code{methodObjPredict} method to obtain 
predicted values.}
}}

\section{Methods}{

  \describe{
    \item{fitObject}{: Extracts regression step. }
    \item{model}{: Retrieve model. }
    \item{solver}{: Retrieve regression method name. }
    \item{solverArgs}{: Retrieve arguments to be sent to regression method. }
    \item{solverArgs(object)<-}{: Set arguments to be sent to regression method. }
    \item{predictor}{: Retrieve prediction method name. }
    \item{predictorArgs}{: Retrieve arguments to be sent to prediction method. }
    \item{predictorArgs(object)<-}{: Set arguments to be sent to prediction method. }
  }
}

\examples{
showClass("modelObjFit")

}
