% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A01_methodObj.R
\docType{methods}
\name{newMethodObj}
\alias{newMethodObj}
\alias{.newMethodObj,ANY,ANY-method}
\alias{.newMethodObj,character,list-method}
\alias{.newMethodObj,function,list-method}
\title{Create an object of class methodObj}
\usage{
.newMethodObj(method, args)

\S4method{.newMethodObj}{ANY,ANY}(method, args)

\S4method{.newMethodObj}{character,list}(method, args)

\S4method{.newMethodObj}{`function`,list}(method, args)
}
\arguments{
\item{method}{A character providing the name of the function or the function
itself.}

\item{args}{A list of input arguments to be passed to the function when 
called.}
}
\value{
An object of class \code{methodObj}.
}
\description{
Creates an object of class methodObj that holds the regression or prediction
  method and the necessary inputs.
}
\keyword{internal}
