% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A02_methodObjPredict.R
\name{newMethodObjPredict}
\alias{newMethodObjPredict}
\title{Create an object of class methodObjPredict}
\usage{
.newMethodObjPredict(args, ...)
}
\arguments{
\item{args}{A list of input arguments}

\item{method}{A character name or the function for making predictions}
}
\value{
An object of class methodObjPredict
}
\description{
Creates an object of class methodObjPredict
}
\keyword{internal}
