% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A03_methodObjSolver.R
\name{newMethodObjSolver}
\alias{newMethodObjSolver}
\title{Create an object of class methodObjSolver}
\usage{
.newMethodObjSolver(args, ...)
}
\arguments{
\item{args}{A list of input arguments}

\item{method}{A character name or the function for regression}
}
\value{
An object of class methodObjSolver
}
\description{
Creates an object of class methodObjSolver
}
\keyword{internal}
