% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A04_methodObjSolverFormula.R
\name{newMethodObjSolverFormula}
\alias{newMethodObjSolverFormula}
\title{Create an object of class methodObjSolverFormula}
\usage{
.newMethodObjSolverFormula(args, ...)
}
\arguments{
\item{args}{A list of input arguments}
}
\value{
An object of class methodObjSolverFormula
}
\description{
Creates an object of class methodObjSolverFormula
}
\keyword{internal}
