% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A05_methodObjSolverXY.R
\name{newMethodObjSolverXY}
\alias{newMethodObjSolverXY}
\title{Create an object of class methodObjSolverXY}
\usage{
.newMethodObjSolverXY(args, ...)
}
\arguments{
\item{args}{A list of input arguments}
}
\value{
An object of class methodObjSolverXY
}
\description{
Creates an object of class methodObjSolverXY
}
\keyword{internal}
