% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelObj.R, R/modelObjFit.R
\name{predictorArgs}
\alias{predictorArgs}
\alias{predictorArgs,modelObj-method}
\alias{predictorArgs<-}
\alias{predictorArgs<-,ANY,ANY-method}
\alias{predictorArgs<-,modelObj,list-method}
\alias{predictorArgs,modelObjFit-method}
\title{Retrieve Predictor Arguments}
\usage{
predictorArgs(object, ...)

\S4method{predictorArgs}{modelObj}(object, ...)

predictorArgs(object) <- value

\S4method{predictorArgs}{ANY,ANY}(object) <- value

\S4method{predictorArgs}{modelObj,list}(object) <- value

\S4method{predictorArgs}{modelObjFit}(object, ...)
}
\arguments{
\item{object}{A modelObj object}

\item{...}{ignored}

\item{value}{List to be stored in args}
}
\value{
A list
}
\description{
Retrieves the arguments that are to be passed to the prediction method
  when called.
}
