% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelObj.R, R/modelObjFit.R
\name{solver}
\alias{solver}
\alias{solver,ANY-method}
\alias{solver,modelObj-method}
\alias{solver,modelObjFit-method}
\title{Retrieve Solver Method}
\usage{
solver(object, ...)

\S4method{solver}{ANY}(object, ...)

\S4method{solver}{modelObj}(object, ...)

\S4method{solver}{modelObjFit}(object, ...)
}
\arguments{
\item{object}{A modelObj object}

\item{...}{ignored}
}
\value{
An object of class character or function
}
\description{
Retrieves method for regression analysis
}
