% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelObj.R, R/modelObjFit.R
\name{solverArgs}
\alias{solverArgs}
\alias{solverArgs,ANY-method}
\alias{solverArgs,modelObj-method}
\alias{solverArgs<-}
\alias{solverArgs<-,ANY,ANY-method}
\alias{solverArgs<-,modelObj,list-method}
\alias{solverArgs,modelObjFit-method}
\title{Retrieve Solver Arguments}
\usage{
solverArgs(object, ...)

\S4method{solverArgs}{ANY}(object, ...)

\S4method{solverArgs}{modelObj}(object, ...)

solverArgs(object) <- value

\S4method{solverArgs}{ANY,ANY}(object) <- value

\S4method{solverArgs}{modelObj,list}(object) <- value

\S4method{solverArgs}{modelObjFit}(object, ...)
}
\arguments{
\item{object}{A modelObj object}

\item{...}{ignored}

\item{value}{List to be stored in args}
}
\value{
A list
}
\description{
Retrieves the arguments that are to be passed to the regression method
  when called.
}
