% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnMsg.R
\name{warnMsg}
\alias{warnMsg}
\title{Generate Warning that Method Does not Exist}
\usage{
warnMsg(x, cx)
}
\arguments{
\item{x}{character name of method requested}

\item{cx}{character name of the class of the object for which the method
was requested}
}
\description{
Function generates and prints warning message when a method is not defined
  for a regression method.
}
\keyword{internal}
