% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A06_modelObj.R, R/A09_modelObjFit.R
\docType{methods}
\name{model}
\alias{model}
\alias{model,modelObjFit-method}
\title{Retrieve model}
\usage{
model(object, ...)

\S4method{model}{modelObjFit}(object, ...)
}
\arguments{
\item{object}{A modelObj object}

\item{...}{ignored}
}
\value{
The formula for the regression
}
\description{
Retrieves model from modelObj
}
