% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelObj.R, R/modelObjFit.R
\name{predictor}
\alias{predictor}
\alias{predictor,modelObj-method}
\alias{predictor,modelObjFit-method}
\title{Retrieve Prediction Method}
\usage{
predictor(object, ...)

\S4method{predictor}{modelObj}(object, ...)

\S4method{predictor}{modelObj}(object, ...)

\S4method{predictor}{modelObjFit}(object, ...)
}
\arguments{
\item{object}{A modelObj object}

\item{...}{ignored}
}
\value{
An object of class character or function
}
\description{
Retrieves method for prediction analysis
}
