% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A06_modelObj.R, R/A09_modelObjFit.R
\docType{methods}
\name{predictorArgs}
\alias{predictorArgs}
\alias{predictorArgs<-}
\alias{predictorArgs,modelObjFit-method}
\title{Retrieve Predictor Arguments}
\usage{
predictorArgs(object, ...)

predictorArgs(object) <- value

\S4method{predictorArgs}{modelObjFit}(object, ...)
}
\arguments{
\item{object}{A modelObj object}

\item{...}{ignored}

\item{value}{List to be stored in args}
}
\value{
A list
}
\description{
Retrieves the arguments that are to be passed to the prediction method
  when called.
}
