% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A06_modelObj.R, R/A09_modelObjFit.R
\docType{methods}
\name{solver}
\alias{solver}
\alias{solver,modelObjFit-method}
\title{Retrieve Solver Method}
\usage{
solver(object, ...)

\S4method{solver}{modelObjFit}(object, ...)
}
\arguments{
\item{object}{A modelObj object}

\item{...}{ignored}
}
\value{
An object of class character or function
}
\description{
Retrieves method for regression analysis
}
