% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A06_modelObj.R, R/A09_modelObjFit.R
\docType{methods}
\name{solverArgs}
\alias{solverArgs}
\alias{solverArgs<-}
\alias{solverArgs,modelObjFit-method}
\title{Retrieve Solver Arguments}
\usage{
solverArgs(object, ...)

solverArgs(object) <- value

\S4method{solverArgs}{modelObjFit}(object, ...)
}
\arguments{
\item{object}{A modelObj object}

\item{...}{ignored}

\item{value}{List to be stored in args}
}
\value{
A list
}
\description{
Retrieves the arguments that are to be passed to the regression method
  when called.
}
